+++ src/index.css
@@ -318,3 +318,15 @@ h1, h2, h3, h4, h5, h6 {
 .comment {
   width: 100%;
 }
+
+
+.main-nav {
+  display: flex;
+  gap: 10px;
+}
+
+.loading {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
