+++ src/pages/Home.jsx
@@ -6,15 +6,18 @@ import Pagination from "../components/Pagination";
 import ArticlesList from "../components/ArticlesList";
 import axios from "axios";
 import { useEffect, useState } from "react";
+import LoadingScreen from "../components/LoadingScreen";
 
 export default () => {
 
   const [articles, setArticles] = useState([])
+  const [isLoading,setIsLoading] = useState(true)
 
   useEffect(()=>{
     axios.get("https://nc-news-ngma.onrender.com/api/articles?sort_by=votes")
     .then((response)=>{
       setArticles(response.data.articles)
+      setIsLoading(false)
     })
     .catch((err)=>{
       console.log(err);
@@ -22,22 +25,26 @@ export default () => {
   },[])
 
   return (
-<main className="container">
-  <FeaturedArticleLarge article={articles[0]}/>
-  <FeaturedArticleSmall articles={articles.slice(1,3)}/>
-  <div className="row g-5">
-    <div className="col-md-8">
-      <ArticlesList articles={articles.slice(3)}/>
-      <Pagination />
-    </div>
-    <div className="col-md-4">
-      <div className="position-sticky">
-    <SidebarCard />
-    <RecentPosts articles={articles}/>
-      </div>
-    </div>
-  </div>
-</main>
-  )
+    <main className="container">
+      {isLoading ? <LoadingScreen /> : (
+        <>
+          <FeaturedArticleLarge article={articles[0]} />
+          <FeaturedArticleSmall articles={articles.slice(1, 3)} />
+          <div className="row g-5">
+            <div className="col-md-8">
+              <ArticlesList articles={articles.slice(3)} />
+              <Pagination />
+            </div>
+            <div className="col-md-4">
+              <div className="position-sticky">
+                <SidebarCard />
+                <RecentPosts articles={articles} />
+              </div>
+            </div>
+          </div>
+        </>
+      )}
+    </main>
+  );
 }
